fix: AuthRoute 수정 및 순환참조 제거
