+++ frontend/src/App.tsx
@@ -5,7 +5,8 @@ import axios from "axios";
 import AuthRoute from "./components/Auth/AuthRoute/AuthRoute";
 import Header from "./components/Header/Header";
 import useAuthContext from "./hooks/useAuthContext";
-import { PATH, ROUTE } from "./utils/constants/route";
+import { PATH } from "./utils/constants/path";
+import { ROUTE } from "./utils/constants/route";
 
 axios.defaults.baseURL = process.env.SERVER_BASE_URL;
