+++ frontend/src/components/Header/Header.tsx
@@ -3,7 +3,8 @@ import { Link, NavLink } from "react-router-dom";
 import Logo from "../../assets/logo.svg";
 import useAuthContext from "../../hooks/useAuthContext";
 import useModalContext from "../../hooks/useModalContext";
-import { NAV_MENU, PATH } from "../../utils/constants/route";
+import { PATH } from "../../utils/constants/path";
+import { NAV_MENU } from "../../utils/constants/route";
 import GithubOAuth from "../Auth/OAuth/GithubOAuth";
 import Button from "../shared/Button/Button";
 import { FlexCenter } from "../shared/Flexbox/Flexbox";
