+++ frontend/src/utils/constants/path.ts
@@ -1,0 +1,8 @@
+export const PATH = {
+  MAIN: "/",
+  REVIEWER_DETAIL: "/reviewer/:reviewerId",
+  REVIEWER_REGISTER: "/reviewer/register",
+  REVIEW_HISTORY: "/review/history",
+  REVIEW_REQUEST: "/review/request/:reviewerId",
+  REDIRECT_OAUTH: "/redirect/oauth",
+};
